modified adapter code

+++ app/src/main/java/com/abp/driver/fragment/DriverAttendanceDetailFragment.java
@@ -11,6 +11,8 @@ import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
+import android.widget.Toast;
+
 import com.abp.driver.ApiClient.ApiClients;
 import com.abp.driver.Interface.Api;
 import com.abp.driver.R;
@@ -18,7 +20,10 @@ import com.abp.driver.activity.DashboardActivity;
 import com.abp.driver.adapter.DriverAttDetAdapter;
 import com.abp.driver.model.driver.DriverAttendance;
 import com.abp.driver.model.driver.DriverAttendanceList;
+import com.abp.driver.utils.Constant;
 import com.abp.driver.utils.CustomLog;
+import com.abp.driver.utils.SharedPreference;
+
 import java.util.List;
 import butterknife.BindView;
 import butterknife.ButterKnife;
@@ -33,6 +38,10 @@ public class DriverAttendanceDetailFragment extends Fragment {
     private RecyclerView.LayoutManager layoutManager;
     @BindView(R.id.rv_dri_att_det)
     RecyclerView mRecyclerView;
+    private SharedPreference mSharePref;
+    private List<DriverAttendanceList> mList;
+    private DashboardActivity mActivity;
+
     @Nullable
     @Override
     public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
@@ -41,45 +50,67 @@ public class DriverAttendanceDetailFragment extends Fragment {
         View view = inflater.inflate(R.layout.fragment_driver_attendance_detail, container, false );
 
         ub_dri_att_det = ButterKnife.bind(this, view);
-
-        layoutManager = new LinearLayoutManager(getActivity());
-        mRecyclerView.setLayoutManager(layoutManager);
-
-        DriverAttDetAdapter driverAttDetAdapter = new DriverAttDetAdapter();
-        mRecyclerView.setAdapter(driverAttDetAdapter);
-
-        callAttendanceApi();
         init();
        return view;
     }
     private void init() {
-        DashboardActivity mActivity = (DashboardActivity) getActivity();
+        mActivity = (DashboardActivity) getActivity();
         if (mActivity != null) {
             mActivity.setToolbarTitle("Driver Attendance");
         }
+        mSharePref = new SharedPreference(getContext());
+        mList = DriverAttendanceList.listAll(DriverAttendanceList.class);
+        if (mList.size() > 0) {
+            callRecyclerView();
+            if (mActivity.isNetworkAvailable())
+            callAttendanceApi();
+        } else {
+            if (mActivity.isNetworkAvailable())
+                callAttendanceApi();
+            else
+                Toast.makeText(getContext(),"No internet available ",Toast.LENGTH_SHORT).show();
+        }
     }
     private void callAttendanceApi() {
-        String api_key = "abpn";
-        String phone_no = "9005103632";
-        Api api = ApiClients.getApiClients().create(Api.class);
-        Call<DriverAttendance> listCall = api.driverAttendance(api_key, phone_no);
-        listCall.enqueue(new Callback<DriverAttendance>() {
-            @Override
-            public void onResponse(Call<DriverAttendance> call, Response<DriverAttendance> response) {
+        try {
+            String api_key = Constant.API_KEY;
+            String phone_no = mSharePref.getUserPhoneNo();
+            Api api = ApiClients.getApiClients().create(Api.class);
+            Call<DriverAttendance> listCall = api.driverAttendance(api_key, phone_no);
+            listCall.enqueue(new Callback<DriverAttendance>() {
+                @Override
+                public void onResponse(Call<DriverAttendance> call, Response<DriverAttendance> response) {
+                    DriverAttendance driverAttendance = response.body();
+                    if (driverAttendance.getSTATUS().equals(Constant.SUCCESS_CODE)) {
+                        DriverAttendanceList.deleteAll(DriverAttendanceList.class);
+                        for (DriverAttendanceList driverAttendanceList : driverAttendance.getData()) {
+                            driverAttendanceList.save();
+                        }
+                        callRecyclerView();
+                        CustomLog.d("recyclerList", "Responding");
+                    } else {
 
-                DriverAttendance driverAttendance = response.body();
-                DriverAttendanceList.deleteAll(DriverAttendanceList.class);
-                for (DriverAttendanceList driverAttendanceList : driverAttendance.getData()){
-                    driverAttendanceList.save();
+                    }
+
                 }
-                CustomLog.d("recyclerList", "Responding");
+                @Override
+                public void onFailure(Call<DriverAttendance> call, Throwable t) {
+                    CustomLog.d("recyclerList", "Not Responding");
+                }
+            });
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+    }
 
-            }
-            @Override
-            public void onFailure(Call<DriverAttendance> call, Throwable t) {
-                CustomLog.d("recyclerList", "Not Responding");
-            }
-        });
+    private void callRecyclerView() {
+        mList = DriverAttendanceList.listAll(DriverAttendanceList.class);
+        if (mList.size() > 0) {
+            layoutManager = new LinearLayoutManager(getActivity());
+            mRecyclerView.setLayoutManager(layoutManager);
+            DriverAttDetAdapter driverAttDetAdapter = new DriverAttDetAdapter(getContext(),mList);
+            mRecyclerView.setAdapter(driverAttDetAdapter);
+        }
     }
 
     @Override
